requirements.txt
# PreGrade Backend Dependencies

# Image processing
Pillow>=10.0.0
numpy>=1.24.0

# OCR
pytesseract>=0.3.10

# ML (optional, for ONNX inference)
onnx>=1.15.0
onnxruntime>=1.17.0

# Testing
pytest>=7.0.0

# Data (dev/optional)
kagglehub>=0.2.0
